Fix issue #88
2 changes here:

 * Revert "Flush remaining data when dynamic content is delivered"

This reverts commit d09ed3dc622be8d2b17efcd1cda552ee777fe464.
Systematically flushing data is not a good idea.

 * Add "flush" as possible return value of the out/1 function

Some clients (eg. Firefox and google-chome...) do not handle connection
close gracefully when the connection is closed during sending data; they
do not read the server response. So, when the response is important, to
warn the end-user that an error occured for example, we must be fair by
flushing data before closing the connection.