Refactor the reverse proxy to manage it as a "special" appmod
Now, the reverse proxy is written like an appmod and follows the same
workflow. We forge the #urltype{} record and use Arg#arg.state to pass
the revproxy's rule before calling yaws_revproxy:out/1.

For now, the support of persistent connections with the backend servers
is disabled. But it does not affect the status of client connections.
Chunked-transfer encoded requests and responses are supported.

Despite this refactoring, yaws_revproxy is always experimental and need
more tests to be safely used in production environments.