Fix is_revproxy/3 and is_redirect_map/2 in yaws.erl
In these functions, we try to find the longest matching prefix and not
the first anymore. So we can have several revproxy/redirect_map definitions
for which order does not matter.

For example, if we define these revproxy rules (in any order):
  revproxy = /goo    http://www.worldofgoo.com     (1)
  revproxy = /       http://yaws.hyber.org         (2)
  revproxy = /erl    http://www.erlang.org         (3)
  revproxy = /google http://www.google.com         (4)

then
  http://localhost will match the rule (2)
  http://localhost/goo will match the rule (1)
  http://localhost/google/search will match the rule (4)
  http://localhost/erl will match the rule (3)

We also change how the matching is made. Now, the matching is done by
splitting the path on slashes. So, using previous rules, "/google" is the
longest prefix for the path "/google/test" and "/" is the longest prefix
for "/erlang".