Add 'Last-Modified' header only for static files or 304 responses
It is a partial rollback of the commit 89f3ea22.
The browsers seem to use this header to do aggressive cache (using some
strange heuristics), without doing any request to the server. IMHO, it is
a bug in browsers. Nevertheless, this header is useless for dynamic resources.
So we can skip it. When this header is meaningful for user's application, it
could be added explicitly.