fix #168 by checking for pam headers and library
Make configure script check for the pam_appl.h header in multiple
directories, and add the directory in which it's found to CFLAGS. Ensure
the checking takes EXTRAINCLUDE settings into account. Remove
specifications of the /usr/include/pam include path from CFLAGS in the
configure script, and also from c_src/Makefile, as it's no longer necessary
to specify that path explicitly. Add EXTRAINCLUDE to CFLAGS and CPPFLAGS
when pam is enabled (note that adding it to CPPFLAGS is needed so the new
configure checking for pam_appl.h works correctly), and remove the addition
of EXTRAINCLUDE to CFLAGS in c_src/Makefile. Add check for pam library. Fix
checking of the --disable-pam configure option.

Thanks to Antoine Catton for starting the work on these changes.