use ibrowse v3.0.4 for testing
On OS X (and perhaps elsewhere too) there were tons of false errors in the
test output, all coming from the ibrowse client complaining about closed
connections. This made it impossible to know whether the tests were
actually working or not. Instead of using ibrowse master, pin the version
the yaws tests use to 3.0.4, which doesn't exhibit this problem.