Do not set recbuf by default on listening sockets
This patch removes a bug I experienced on FreeBSD during tests on the loopback
interface because of its MTU. So, I don't really know why that happens on
FreeBSD and not on Linux, but because loopback MTU is greater than the recbuf
value (16K on loopback vs 8k in Yaws), if a client sends data, Yaws will read
them very slowly.

Furthermore, there is no good reason to tweak this value by default. And, if
required, it is always possible to configure it.