For rebar, make ibrowse a dep only for Yaws dev mode
For rebar builds, rather than treating ibrowse as a normal dependency even
though it's used only for testing, make rebar.config.script check for the
existence of a .rebar/YAWS_DEV_MODE file and only if present, add ibrowse
as a dependency. This allows projects that use Yaws as a rebar dependency
to avoid having to pull in ibrowse.

(Credit goes to Seth Falcon for this idea; he added something like this to
the Basho webmachine project for a test-only dependency there, and I
borrowed his idea for Yaws.)